[frontend] refactor schemas
