+++ bankme-frontend/src/components/assignor/update-assignor.tsx
@@ -22,13 +22,7 @@ import { Input } from "@/components/ui/input";
 import { Button, buttonVariants } from "@/components/ui/button";
 import { Assignor, updateAssignor } from "@/services/assignor";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-
-const formSchema = z.object({
-  document: z.string().min(1, { message: "document is required" }).max(30),
-  email: z.string().email().max(140),
-  phone: z.string().min(1, { message: "phone is required " }).max(20),
-  name: z.string().min(1, { message: "name is required" }).max(140),
-});
+import { assignorSchema } from "@/schemas/assignor-schema";
 
 interface Props {
   assignor: Assignor;
@@ -36,8 +30,8 @@ interface Props {
 export default function UpdateAssignor({ assignor }: Props) {
   const [open, setOpen] = useState(false);
 
-  const form = useForm<z.infer<typeof formSchema>>({
-    resolver: zodResolver(formSchema),
+  const form = useForm<z.infer<typeof assignorSchema>>({
+    resolver: zodResolver(assignorSchema),
     defaultValues: {
       document: assignor.document,
       email: assignor.email,
@@ -59,7 +53,7 @@ export default function UpdateAssignor({ assignor }: Props) {
     },
   });
 
-  function onSubmit(data: z.infer<typeof formSchema>) {
+  function onSubmit(data: z.infer<typeof assignorSchema>) {
     const { document, email, name, phone } = data;
     updateAssignorMutation({ id: assignor.id, document, email, name, phone });
   }
